pyproject.toml
[project]
name = "project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "google-genai>=1.28.0",
    "jinja2>=3.1.6",
    "langchain[google-genai]>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "pypdf2>=3.0.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "uvicorn>=0.35.0",
    "dotenv>=0.9.9",
    "lxml>=6.0.0",
    "tabulate>=0.9.0",
]
